Resolve string handling for "https://" or "http://" in update_rhsm_conf
Previously, if a proxy URI was entered into the registration page leading with
"https://" or "http://", it would be incorrectly entered into the rhsm.conf.
This patch prevents this by ensuring it is a valid URI, and correctly parsing
the string regardless if it has only a host in the URI or if it leads with the
protocol.